Use cat/gzip instead of keep-latest
For the secondary package file, it was wrong to use keep-latest, as
Modestas Vainius pointed out. With this change, files are just
copied/unpacked, allowing edos-debcheck to correctly determine if the
build-deps can be satisfied.

This change affects the triggers for all non-standard (*-volatile,
*-bpo, *-edu and experimental) suites, which should all use the
aptitude dep resolver (able to handle such complex dep situations) on
the buildds.