Use public.locks as locking area.
Since the central packages table was introduced we used FOR UPDATE
for synchronization between different wanna-build calls.  This
causes a write for every locked row, which is unhelpful for WAL
traffic considering that merge-v3 needs to lock everything.

Let's synchronize by using FOR UPDATE on public.locks instead,
which should ease the WAL load a lot.