Permit architecture qualifiers, and strip them for edos-debcheck
Architecture qualifiers (such as :any and :native) work fine with
current sbuild.  However, edos-debcheck doesn't understand them, so
strip them off in the synthetic Packages files passed to edos-debcheck.

The worst case here should be that packages with "Build-Depends:
foo:any" may be scheduled for building and then fail if "foo" is not
"Multi-Arch: allowed"; this is not too bad, and it usually corresponds
to a maintainer error anyway.