Convert to dose-builddebcheck
edos-debcheck is deprecated, and since we can only make it work by
treating source packages as binary packages for the purpose of
installability checking, it is fundamentally incapable of handling newer
features such as multiarch ":any" dependency annotations properly.
dose-builddebcheck is the modern replacement and does a better job of
this.

Deploying this requires first installing liburi-perl and
dose-builddebcheck.